进行了有限调整。
